+++ spec/models/user_activity_spec.rb
@@ -114,6 +114,32 @@ describe UserActivity do
       )
     end
   end
+  context 'swim activity in meters' do
+    let(:team) { Fabricate(:team, units: 'km') }
+    let(:user) { Fabricate(:user, team: team) }
+    let(:activity) { Fabricate(:swim_activity, user: user) }
+    it 'to_slack' do
+      expect(activity.to_slack).to eq(
+        attachments: [
+          {
+            fallback: "#{activity.name} via #{activity.user.slack_mention}, 1874m 37m 1m58s/100m",
+            title: activity.name,
+            title_link: "https://www.strava.com/activities/#{activity.strava_id}",
+            text: "<@#{activity.user.user_name}> on Tuesday, February 20, 2018 at 10:02 AM",
+            fields: [
+              { title: 'Type', value: 'Swim 🏊', short: true },
+              { title: 'Distance', value: '1874m', short: true },
+              { title: 'Time', value: '37m', short: true },
+              { title: 'Pace', value: '1m58s/100m', short: true }
+            ],
+            author_name: user.athlete.name,
+            author_link: user.athlete.strava_url,
+            author_icon: user.athlete.profile_medium
+          }
+        ]
+      )
+    end
+  end
   context 'maps' do
     context 'without maps' do
       let(:team) { Fabricate(:team, maps: 'off') }